fixed: execution section inspection is fixed refactored: tests and error messages
Resolves: #408
