+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkUnrecognizedSectionInspection.kt
@@ -5,18 +5,15 @@ import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.util.elementType
 import com.intellij.codeInspection.ui.ListEditForm
 import com.jetbrains.snakecharm.SnakemakeBundle
+import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.EXECUTION_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.MODULE_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS
-import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.RULE_OR_CHECKPOINT_SECTION_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.SUBWORKFLOW_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.USE_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.lang.psi.*
 import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes
 import com.jetbrains.snakecharm.inspections.quickfix.AddIgnoredElementQuickFix
-import com.jetbrains.snakecharm.lang.SnakemakeNames.MODULE_KEYWORD
-import com.jetbrains.snakecharm.lang.SnakemakeNames.RULE_KEYWORD
-import com.jetbrains.snakecharm.lang.SnakemakeNames.SUBWORKFLOW_KEYWORD
-import com.jetbrains.snakecharm.lang.SnakemakeNames.USE_KEYWORD
+import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_RUN
 import javax.swing.JComponent
 
 class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
@@ -56,16 +53,14 @@ class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
             val sectionKeyword = argsSection.sectionKeyword
 
             if (sectionNamePsi != null && sectionKeyword != null && sectionKeyword !in setOfValidNames
-                && sectionKeyword !in ignoredItems
+                && sectionKeyword !in ignoredItems && !(argsSection.getParentRuleOrCheckPoint() is SmkUse
+                        && sectionKeyword in (EXECUTION_SECTIONS_KEYWORDS + SECTION_RUN))
             ) {
-                val appropriateSection = getSectionBySubsection(sectionKeyword)
-                if (appropriateSection == null) {
-                    registerProblem(
-                        sectionNamePsi,
-                        SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword),
-                        AddIgnoredElementQuickFix(sectionKeyword)
-                    )
-                }
+                registerProblem(
+                    sectionNamePsi,
+                    SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword),
+                    AddIgnoredElementQuickFix(sectionKeyword)
+                )
             }
         }
 
@@ -78,15 +73,5 @@ class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
 
     override fun createOptionsPanel(): JComponent? =
         ListEditForm(SnakemakeBundle.message("INSP.NAME.section.unrecognized.ignored"), ignoredItems).contentPanel
-
-    companion object {
-        fun getSectionBySubsection(name: String) = when (name) {
-            in RULE_OR_CHECKPOINT_SECTION_KEYWORDS -> RULE_KEYWORD
-            in SUBWORKFLOW_SECTIONS_KEYWORDS -> SUBWORKFLOW_KEYWORD
-            in MODULE_SECTIONS_KEYWORDS -> MODULE_KEYWORD
-            in USE_SECTIONS_KEYWORDS -> USE_KEYWORD
-            else -> null
-        }
-    }
 }
