+++ src/main/resources/SnakemakeBundle.properties
@@ -14,8 +14,8 @@ INSP.NAME.section.unrecognized.message=Section ''{0}'' isn''t recognized by Snak
 INSP.NAME.section.unrecognized.ignored=Ignored Sections
 INSP.NAME.section.unrecognized.ignored.add=Ignore an unrecognized section ''{0}''
 
-# SmkUnexpectedSectionInspection
-INSP.NAME.section.unexpected=Section ''{0}'' can''t be used in ''{1}'' but can be in ''{2}''
+# SmkExecutionSubsectionInUseSectionInspection
+INSP.NAME.unexpected.execution.section=Execution sections can't be overridden in 'use rule'
 
 # SmkUnusedLogFileInspection
 INSP.NAME.unused.log.section=Looks like a log file won't be created, because it is not referenced from 'shell' or 'run' sections
