+++ src/main/resources/inspectionDescriptions/SmkExecutionSubsectionInUseSectionInspection.html
@@ -1,0 +1,6 @@
+<html>
+<body>
+This inspection highlights the sections which can't be overridden in 'use rule' ('shell', 'script',
+'notebook', 'cwl' and 'run' sections).
+</body>
+</html>
