--- src/main/resources/inspectionDescriptions/SmkUnexpectedSectionInspection.html
@@ -1,6 +1,0 @@
-<html>
-<body>
-This inspection highlights the subsections which can't be used in current section,
-but can be used in other sections.
-</body>
-</html>
