+++ src/test/resources/features/highlighting/inspections/unrecognized_section_inspection.feature
@@ -22,22 +22,22 @@ Feature: Inspection if section isn't recognized by SnakeCharm
       | subworkflow |
       | checkpoint  |
 
-    Scenario Outline: When quick fix is, applied check no warning
-      Given a snakemake project
-      Given I open a file "foo.smk" with text
+  Scenario Outline: When quick fix is, applied check no warning
+    Given a snakemake project
+    Given I open a file "foo.smk" with text
       """
       <rule_like>:
           unknown_section: ""
       """
-      And SmkUnrecognizedSectionInspection inspection is enabled
-      And I emulate quick fix apply: ignore unresolved item 'unknown_section'
-      Then I expect no inspection weak warnings
-      When I check highlighting weak warnings
-      Examples:
-        | rule_like   |
-        | rule        |
-        | subworkflow |
-        | checkpoint  |
+    And SmkUnrecognizedSectionInspection inspection is enabled
+    And I emulate quick fix apply: ignore unresolved item 'unknown_section'
+    Then I expect no inspection weak warnings
+    When I check highlighting weak warnings
+    Examples:
+      | rule_like   |
+      | rule        |
+      | subworkflow |
+      | checkpoint  |
 
   Scenario Outline: When quick fix is applied, check such element in ignored list
     Given a snakemake project
@@ -59,3 +59,25 @@ Feature: Inspection if section isn't recognized by SnakeCharm
       | rule        |
       | subworkflow |
       | checkpoint  |
+
+  Scenario: No weak warnings on execution sections in 'use rule'
+    Given a snakemake project
+    Given I open a file "foo.smk" with text
+      """
+      rule A:
+          threads: 4
+
+      use rule A as B with:
+          run: ""
+          shell: ""
+          notebook: ""
+          input: "file1"
+          script: ""
+          cwl: ""
+          wrapper: ""
+          output: "file2"
+      """
+    And SmkUnrecognizedSectionInspection inspection is enabled
+    And I emulate quick fix apply: ignore unresolved item 'unknown_section'
+    Then I expect no inspection weak warnings
+    When I check highlighting weak warnings
